+++ torch_frame/typing.py
@@ -1,7 +1,14 @@
+from enum import Enum
 from typing import List, Union
 
 import pandas as pd
 from torch import Tensor
+
+
+class TaskType(Enum):
+    REGRESSION = 'regression'
+    MULTICLS_CLASSIFICATION = 'multicls_classification'
+
 
 Series = pd.Series
 DataFrame = pd.DataFrame